switch default handling of py floats as rs floats

+++ parsons/databases/redshift/constants.py
@@ -1,6 +1,6 @@
 COL_NAME_MAX_LEN = 120
 
-DECIMAL = 'decimal'
+DECIMAL = 'float'
 
 REPLACE_CHARS = {" ": ""}
